+++ node/src/chain_specs/mod.rs
@@ -18,7 +18,10 @@
 #![allow(dead_code)]
 use cumulus_primitives_core::ParaId;
 use hex_literal::hex;
-use manta_primitives::{constants, AccountId, AuraId, Balance, Signature};
+use manta_primitives::{
+	constants,
+	types::{AccountId, AuraId, Balance, Signature},
+};
 use sc_chain_spec::{ChainSpecExtension, ChainSpecGroup};
 use sc_service::{ChainType, Properties};
 use serde::{Deserialize, Serialize};