+++ runtime/primitives/src/lib.rs
@@ -19,61 +19,43 @@
 #![allow(clippy::upper_case_acronyms)]
 #![cfg_attr(not(feature = "std"), no_std)]
 
+pub mod assets;
+pub mod constants;
+pub mod types;
+pub mod xcm;
+
 #[cfg(feature = "std")]
 pub mod helpers;
 
-mod assets;
-pub mod constants;
-mod xcm;
-pub use crate::{
-	assets::{
-		AssetIdLocationConvert, AssetIdLocationGetter, AssetLocation, AssetRegistarMetadata,
-		AssetStorageMetadata, UnitsToWeightRatio,
-	},
-	xcm::{
-		AccountIdToMultiLocation, FirstAssetTrader, IsNativeConcrete, MultiNativeAsset,
-		XcmFeesToAccount,
-	},
-};
-pub use constants::*;
-
-use sp_runtime::traits::{BlakeTwo256, IdentifyAccount, Verify};
-
-/// An index to a block.
-pub type BlockNumber = u32;
-
-/// Alias to 512-bit hash when used in the context of a transaction signature on the chain.
-pub type Signature = sp_runtime::MultiSignature;
-
-/// Some way of identifying an account on the chain. We intentionally make it equivalent
-/// to the public key of our transaction signing scheme.
-pub type AccountId = <<Signature as Verify>::Signer as IdentifyAccount>::AccountId;
-
-/// The type for looking up accounts. We don't expect more than 4 billion of them, but you
-/// never know...
-pub type AccountIndex = u32;
-
-/// Balance of an account.
-pub type Balance = u128;
-
-/// Index of a transaction in the chain.
-pub type Index = u32;
-
-/// A hash of some data used by the chain.
-pub type Hash = sp_core::H256;
-
-/// Block header type as expected by this runtime.
-pub type Header = sp_runtime::generic::Header<BlockNumber, BlakeTwo256>;
-pub type Block = sp_runtime::generic::Block<Header, sp_runtime::OpaqueExtrinsic>;
-
-/// Digest item type.
-pub type DigestItem = sp_runtime::generic::DigestItem;
-
-// Aura consensus authority.
-pub type AuraId = sp_consensus_aura::sr25519::AuthorityId;
-
-// Moment
-pub type Moment = u64;
-
-// AssetId
-pub type AssetId = u32;
+// From https://github.com/paritytech/polkadot/pull/4332/files?diff=unified&w=1 @ runtime/common/src/lib.rs
+/// Macro to set a value (e.g. when using the `parameter_types` macro) to either a production value
+/// or to an environment variable or testing value (in case the `fast-runtime` feature is selected).
+/// Note that the environment variable is evaluated _at compile time_.
+///
+/// Usage:
+/// ```Rust
+/// parameter_types! {
+/// 	// Note that the env variable version parameter cannot be const.
+/// 	pub LaunchPeriod: BlockNumber = prod_or_fast!(7 * DAYS, 1, "KSM_LAUNCH_PERIOD");
+/// 	pub const VotingPeriod: BlockNumber = prod_or_fast!(7 * DAYS, 1 * MINUTES);
+/// }
+#[macro_export]
+macro_rules! prod_or_fast {
+	($prod:expr, $test:expr) => {
+		if cfg!(feature = "fast-runtime") {
+			$test
+		} else {
+			$prod
+		}
+	};
+	($prod:expr, $test:expr, $env:expr) => {
+		if cfg!(feature = "fast-runtime") {
+			core::option_env!($env)
+				.map(|s| s.parse().ok())
+				.flatten()
+				.unwrap_or($test)
+		} else {
+			$prod
+		}
+	};
+}